Refactoring: divide MasgitoffIo unto multiple classes to improve readablity
