+++ src/test/java/org/mastodon/mamut/collaboration/io/MasgitoffIoTest.java
@@ -31,7 +31,7 @@ public class MasgitoffIoTest
 			final File file = new File( "tmp.masgitoff" );
 			if ( file.isDirectory() )
 				FileUtils.deleteDirectory( file );
-			MasgitoffIo.writeMasgitoff( projectModel.getModel(), file, new MasgitoffIo.MasgitoffIds( graph ) );
+			MasgitoffIo.writeMasgitoff( projectModel.getModel(), file, new MasgitoffIds( graph ) );
 			final ModelGraph newGraph = MasgitoffIo.readMasgitoff( file );
 			ModelAsserts.assertGraphEquals( graph, newGraph );
 		}
